Modularize Contextual Command Bar
