+++ FluentUI.Demo/build.gradle
@@ -15,8 +15,8 @@ android {
         applicationId 'com.microsoft.fluentuidemo'
         minSdkVersion 21
         targetSdkVersion 28
-        versionCode 12
-        versionName '0.0.12'
+        versionCode 13
+        versionName '0.0.13'
     }
     buildTypes {
         release {
@@ -60,6 +60,7 @@ dependencies {
     implementation project(':fluentui_listitem')
     implementation project(':fluentui_drawer')
     implementation project(':fluentui_tablayout')
+    implementation project(':fluentui_ccb')
     implementation project(':fluentui_others')
     implementation 'com.squareup.picasso:picasso:2.71828'
     implementation 'com.github.bumptech.glide:glide:4.8.0'
