+++ config.gradle
@@ -15,5 +15,6 @@ project.ext.fluentui_core_versionid='0.0.1'
 project.ext.fluentui_listitem_versionid='0.0.1'
 project.ext.fluentui_tablayout_versionid='0.0.1'
 project.ext.fluentui_drawer_versionid='0.0.1'
-project.ext.fluentui_others_versionid='0.0.1'
-project.ext.FluentUI_versionid='0.0.12'
+project.ext.fluentui_ccb_versionid='0.0.1'
+project.ext.fluentui_others_versionid='0.0.2'
+project.ext.FluentUI_versionid='0.0.13'
