+++ fluentui_ccb/src/main/java/com/microsoft/fluentui/contextualcommandbar/ContextualCommandBar.kt
@@ -19,14 +19,15 @@ import android.view.ViewGroup.LayoutParams.MATCH_PARENT
 import android.view.ViewGroup.LayoutParams.WRAP_CONTENT
 import android.widget.FrameLayout
 import android.widget.ImageView
-import com.microsoft.fluentui.R
+import com.microsoft.fluentui.ccb.R
+import com.microsoft.fluentui.theming.FluentUIContextThemeWrapper
 import com.microsoft.fluentui.util.isVisible
 
 class ContextualCommandBar @JvmOverloads constructor(
         context: Context,
         attrs: AttributeSet? = null,
         defStyleAttr: Int = 0
-) : FrameLayout(context, attrs, defStyleAttr) {
+) : FrameLayout(FluentUIContextThemeWrapper(context,R.style.Theme_FluentUI_ContextualCommandBar), attrs, defStyleAttr) {
 
     private var dismissButtonContainer: ViewGroup? = null
     private var commandItemAdapter: CommandItemAdapter
@@ -67,8 +68,8 @@ class ContextualCommandBar @JvmOverloads constructor(
         commandItemAdapter = CommandItemAdapter(
                 CommandItemAdapter.CommandListOptions(groupSpace, itemSpace)
         )
-        commandItemRecyclerView = RecyclerView(context).apply {
-            layoutManager = LinearLayoutManager(context, LinearLayoutManager.HORIZONTAL, false)
+        commandItemRecyclerView = RecyclerView(this.context).apply {
+            layoutManager = LinearLayoutManager(this.context, LinearLayoutManager.HORIZONTAL, false)
             adapter = commandItemAdapter
             overScrollMode = View.OVER_SCROLL_NEVER
         }
