+++ fluentui_ccb/src/main/res/color/contextual_command_bar_icon_tint.xml
