+++ fluentui_ccb/src/main/res/drawable-ldrtl/contextual_command_bar_dismiss_button_divider_start_background.xml
