+++ fluentui_ccb/src/main/res/drawable-ldrtl/contextual_command_bar_end_item_background.xml
