+++ fluentui_ccb/src/main/res/drawable/contextual_command_bar_center_item_background.xml
