+++ fluentui_ccb/src/main/res/drawable/contextual_command_bar_dismiss_button_divider_end_background.xml
