+++ fluentui_ccb/src/main/res/drawable/contextual_command_bar_single_item_background.xml
