+++ fluentui_ccb/src/main/res/drawable/contextual_command_bar_start_item_background.xml
