+++ fluentui_ccb/src/main/res/layout/view_command_item.xml
