+++ fluentui_ccb/src/main/res/layout/view_dismiss_command_item.xml
