+++ fluentui_others/src/main/res/values/attrs.xml
@@ -51,16 +51,6 @@
     <attr name="fluentuiCalendarSelectedColor" format="reference|color"/>
     <attr name="fluentuiCalendarDayTodayBackgroundColor" format="reference|color"/>
 
-    <!--ContextualCommandBar-->
-    <attr name="fluentuiContextualCommandBarBackgroundColor" format="reference|color"/>
-    <attr name="fluentuiContextualCommandBarBackgroundColorPressed" format="reference|color"/>
-    <attr name="fluentuiContextualCommandBarBackgroundColorSelected" format="reference|color"/>
-    <attr name="fluentuiContextualCommandBarIconTint" format="reference|color"/>
-    <attr name="fluentuiContextualCommandBarIconTintDisabled" format="reference|color"/>
-    <attr name="fluentuiContextualCommandBarIconTintSelected" format="reference|color"/>
-    <attr name="fluentuiContextualCommandBarDismissBackgroundColor" format="reference|color"/>
-    <attr name="fluentuiContextualCommandBarDismissIconTintColor" format="reference|color"/>
-
     <!--day selector-->
     <attr name="fluentuiCalendarDayTextActiveColor" format="reference|color"/>
     <attr name="fluentuiCalendarDayTextActiveCheckedColor" format="reference|color"/>
