+++ fluentui_others/src/main/res/values/themes.xml
@@ -66,16 +66,6 @@
         <item name="fluentuiCalendarSelectedColor">?attr/fluentuiForegroundSelectedColor</item>
         <item name="fluentuiCalendarDayTodayBackgroundColor">?attr/fluentuiColorPrimaryLighter</item>
 
-        <!--ContextualCommandBar-->
-        <item name="fluentuiContextualCommandBarBackgroundColor">@color/fluentui_gray_50</item>
-        <item name="fluentuiContextualCommandBarBackgroundColorPressed">@color/fluentui_gray_100</item>
-        <item name="fluentuiContextualCommandBarBackgroundColorSelected">@color/fluentui_communication_tint_40</item>
-        <item name="fluentuiContextualCommandBarIconTint">@color/fluentui_gray_900</item>
-        <item name="fluentuiContextualCommandBarIconTintDisabled">@color/fluentui_gray_300</item>
-        <item name="fluentuiContextualCommandBarIconTintSelected">@color/fluentui_communication_blue</item>
-        <item name="fluentuiContextualCommandBarDismissBackgroundColor">#FFFFFF</item>
-        <item name="fluentuiContextualCommandBarDismissIconTintColor">@color/fluentui_gray_900</item>
-
         <!--day selector-->
         <item name="fluentuiCalendarDayTextActiveColor">?attr/fluentuiForegroundSelectedColor</item>
         <item name="fluentuiCalendarDayTextActiveCheckedColor">?attr/fluentuiForegroundOnPrimaryColor</item>
