+++ settings.gradle
@@ -5,4 +5,4 @@ include ':fluentui_drawer'
 include ':fluentui_tablayout'
 include ':fluentui_others'
 include ':FluentUI'
-
+include ':fluentui_ccb'
